group country-locales
