--- Peaqevcore/locale/types/no_tensio.py
@@ -1,13 +1,0 @@
-
-from custom_components.peaqev.peaqservice.util.constants import (
-    QUERYTYPE_AVERAGEOFTHREEDAYS,
-    QUERYTYPE_AVERAGEOFTHREEDAYS_MIN
-)
-from ..locale_model import Locale_Type
-from dataclasses import dataclass
-
-@dataclass(frozen=True)
-class NO_Tensio(Locale_Type):
-    observed_peak = QUERYTYPE_AVERAGEOFTHREEDAYS_MIN
-    charged_peak = QUERYTYPE_AVERAGEOFTHREEDAYS
-
