--- Peaqevcore/locale/types/se_skovde.py
@@ -1,28 +1,0 @@
-from custom_components.peaqev.peaqservice.util.constants import (
-    QUERYTYPE_MAX_NOV_MAR_MON_FRI_06_22
-)
-from dataclasses import dataclass
-from ..locale_model import Locale_Type
-
-@dataclass(frozen=True)
-class SE_Skovde(Locale_Type):
-    
-    observed_peak = QUERYTYPE_MAX_NOV_MAR_MON_FRI_06_22
-    charged_peak = QUERYTYPE_MAX_NOV_MAR_MON_FRI_06_22
-    free_charge_pattern = [
-        {
-            "M": [11, 12, 1, 2, 3],
-            "D": [5, 6],
-            "H": [22, 23, 0, 1, 2, 3, 4, 5]
-        },
-        {
-            "M": [4, 5, 6, 7, 8, 9, 10],
-            "D": [0, 1, 2, 3, 4, 5, 6],
-            "H": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23]
-        }
-    ]
-     
-#November-Mars, vardagar (mån-fre) 06-22
-#single peak i denna period månadsvis.
-
-#ticket 22
